[FRONTEND] feature(ProjectsList): criar componente e remover erro de import do terminal
